Add scan-type(audio,video,image,all) to scanner
Bug-AGL: SPEC-3096

User can pass 'types' as a request property with 'audio','video',
'image','all' values as a single 'string' value or as an array of
'string' values along 'media_result','subscribe','unsubscribe'
verbs to scan or register the events with those scan-types.
If the user doesn't provide the 'types' request property, the API
will function as before with 'audio' and 'video' scan-types.

Change-Id: I8ff49f89e2696f600dfb2935403d8b0c8246c07d
Signed-off-by: Farshid Monhaseri <monhaseri.f@gmail.com>